Move files and views out of the 'main' package into the project root.
I'm going to try the suggestion of putting the project wide stuff in the root project dir in one module. I think it's an interesting idea and want to see how it feels. So, I've created a subscribeme.py module to hold project-wide views and forms and anything else that I can't justify putting into an app.